Refactored the perfomance structure
